Time: 280 ms (39.36%), Space: 121.7 MB (57.30%) - LeetHub

+++ 0295-find-median-from-data-stream/0295-find-median-from-data-stream.cpp
@@ -24,7 +24,7 @@ public:
             return (small_values.top() + big_values.top())/2.0;
         }
         else{
-            return small_values.top();
+            return (float)small_values.top();
         }
     }
 };
